Create scene.cpp in src
